requirements.txt
yfinance
pandas
numpy
sqlalchemy
asyncpg
psycopg2-binary
fastapi
uvicorn
python-dotenv
beautifulsoup4
chromadb
sentence-transformers
ta